pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mojo-bindgen"
version = "0.1.0"
description = "Generate Mojo FFI bindings from C headers using libclang"
readme = "README.md"
license = "MIT"
requires-python = ">=3.14"
authors = [
    { name = "Mohamed Mabrouk", email = "mohamed.mabrouk@rwth-aachen.de" },
]
keywords = ["mojo", "ffi", "bindings", "libclang", "c", "codegen"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: C",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Code Generators",
    "Typing :: Typed",
]
dependencies = [
    "libclang>=18.1.1,<19",
    "rich",
    "typer",
]

[project.urls]
Homepage = "https://github.com/MoSafi2/mojo_bindgen"
Repository = "https://github.com/MoSafi2/mojo_bindgen"
Issues = "https://github.com/MoSafi2/mojo_bindgen/issues"
Documentation = "https://github.com/MoSafi2/mojo_bindgen/blob/main/README.md"
Changelog = "https://github.com/MoSafi2/mojo_bindgen/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "build>=1.4",
    "pyright>=1.1.400",
    "pytest>=9",
    "ruff>=0.8",
]

[project.scripts]
mojo-bindgen = "mojo_bindgen.cli:main"

[tool.hatch.build.targets.sdist]
include = [
    "/LICENSE",
    "/README.md",
    "/CHANGELOG.md",
    "/mojo_bindgen",
    "/pyproject.toml",
]

[tool.hatch.build.targets.wheel]
packages = ["mojo_bindgen"]

[tool.ruff]
line-length = 100
# Ruff supports up to py313 grammar; project runtime is 3.14+.
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "W"]
# E501: long docstrings; UP007: Union[] aliases for recursive IR types are intentional.
ignore = ["E501", "UP007"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501"]

[tool.pyright]
# `libclang` Python bindings ship without accurate stubs; typecheck the rest of the package.
include = ["mojo_bindgen"]
exclude = ["mojo_bindgen/parsing"]
# Align with the oldest supported checker grammar; runtime is 3.14+.
pythonVersion = "3.13"
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = true
